+++ storm-core/src/clj/backtype/storm/config.clj
@@ -94,6 +94,13 @@
       (if (is-absolute-path? path) path (str storm-home file-path-separator path))
       (str storm-home file-path-separator "storm-local"))))
 
+(defn absolute-healthcheck-dir [conf]
+  (let [storm-home (System/getProperty "storm.home")
+        path (conf STORM-HEALTH-CHECK-DIR)]
+    (if path
+      (if (is-absolute-path? path) path (str storm-home file-path-separator path))
+      (str storm-home file-path-separator "healthchecks"))))
+
 (defn master-local-dir
   [conf]
   (let [ret (str (absolute-storm-local-dir conf) file-path-separator "nimbus")]
@@ -273,6 +280,11 @@
   [conf id]
   (LocalState. (worker-heartbeats-root conf id)))
 
+(defn override-login-config-with-system-property [conf]
+  (if-let [login_conf_file (System/getProperty "java.security.auth.login.config")]
+    (assoc conf "java.security.auth.login.config" login_conf_file)
+    conf))
+
 (defn get-topo-logs-users
   [topology-conf]
   (sort (distinct (remove nil?
@@ -286,3 +298,4 @@
                     (concat
                       (topology-conf LOGS-GROUPS)
                       (topology-conf TOPOLOGY-GROUPS))))))
+