+++ storm-core/src/clj/backtype/storm/messaging/local.clj
@@ -14,76 +14,10 @@
 ;; See the License for the specific language governing permissions and
 ;; limitations under the License.
 (ns backtype.storm.messaging.local
-  (:refer-clojure :exclude [send])
-  (:use [backtype.storm log util])
-  (:import [backtype.storm.messaging IContext IConnection TaskMessage])
-  (:import [backtype.storm.grouping Load])
-  (:import [java.util.concurrent LinkedBlockingQueue])
-  (:import [java.util Map Iterator Collection])
-  (:import [java.util Iterator ArrayList])
-  (:gen-class))
-
-(defn update-load! [cached-task->load lock task->load]
-  (locking lock
-    (swap! cached-task->load merge task->load)))
-
-(defn add-queue! [queues-map lock storm-id port]
-  (let [id (str storm-id "-" port)]
-    (locking lock
-      (when-not (contains? @queues-map id)
-        (swap! queues-map assoc id (LinkedBlockingQueue.))))
-    (@queues-map id)))
-
-(deftype LocalConnection [storm-id port queues-map lock queue task->load]
-  IConnection
-  (^Iterator recv [this ^int flags ^int clientId]
-    (when-not queue
-      (throw (IllegalArgumentException. "Cannot receive on this socket")))
-    (let [ret (ArrayList.)
-          msg (if (= flags 1) (.poll queue) (.take queue))]
-      (if msg
-        (do 
-          (.add ret msg)
-          (.iterator ret))
-        nil)))
-  (^void send [this ^int taskId ^bytes payload]
-    (let [send-queue (add-queue! queues-map lock storm-id port)]
-      (.put send-queue (TaskMessage. taskId payload))
-      ))
-  (^void send [this ^Iterator iter]
-    (let [send-queue (add-queue! queues-map lock storm-id port)]
-      (while (.hasNext iter) 
-         (.put send-queue (.next iter)))
-      ))
-  (^void sendLoadMetrics [this ^Map taskToLoad]
-    (update-load! task->load lock taskToLoad))
-  (^Map getLoad [this ^Collection tasks]
-    (locking lock
-      (into {}
-        (for [task tasks
-              :let [load (.get @task->load task)]
-              :when (not-nil? load)]
-          ;; for now we are ignoring the connection load locally
-          [task (Load. true load 0.0)]))))
-  (^void close [this]))
-
-
-(deftype LocalContext [^{:unsynchronized-mutable true} queues-map
-                       ^{:unsynchronized-mutable true} lock
-                       ^{:unsynchronized-mutable true} task->load]
-  IContext
-  (^void prepare [this ^Map storm-conf]
-    (set! queues-map (atom {}))
-    (set! task->load (atom {}))
-    (set! lock (Object.)))
-  (^IConnection bind [this ^String storm-id ^int port]
-    (LocalConnection. storm-id port queues-map lock (add-queue! queues-map lock storm-id port) task->load))
-  (^IConnection connect [this ^String storm-id ^String host ^int port]
-    (LocalConnection. storm-id port queues-map lock nil task->load))
-  (^void term [this]
-    ))
+  (:import [backtype.storm.messaging IContext])
+  (:import [backtype.storm.messaging.local Context]))
 
 (defn mk-context [] 
-  (let [context  (LocalContext. nil nil nil)]
+  (let [context  (Context.)]
     (.prepare ^IContext context nil)
     context))